Fixed the Longest river length issue
It was displaying 1 more than the crrent longest river length. I forgot to take into consideration that the first element represents the line where the river starts at. A simple -1 solved that. Another problem though is that there is no scrolling, so if the string entered is large; most of it will be cut off.